requirements.txt
## Core dependencies
pinecone-client>=3.0.0
pinecone[grpc]
langchain>=0.1.0
langchain-community>=0.0.10
langchain-pinecone>=0.0.3

## Embeddings & ML
numpy>=1.24.0
scikit-learn>=1.3.0
flashrank>=0.2.0
model2vec>=0.1.0

## NLP
nltk>=3.8.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

## Data handling
pandas>=2.0.0
requests>=2.31.0

## Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

## Environment
python-dotenv>=1.0.0
tqdm>=4.66.0
